# perf(sql): speed up first/last aggregate functions for string columns