--- a/core/src/main/java/io/questdb/cairo/TableReaderRecord.java
+++ b/core/src/main/java/io/questdb/cairo/TableReaderRecord.java
@@ -30,6 +30,7 @@ import io.questdb.std.Long128;
 import io.questdb.std.Long256;
 import io.questdb.std.Rows;
 import io.questdb.std.str.CharSink;
+import io.questdb.std.str.DirectSequence;
 import io.questdb.std.str.Sinkable;
 import org.jetbrains.annotations.NotNull;
 
@@ -95,6 +96,18 @@ public class TableReaderRecord implements Record, Sinkable {
                 TableReader.getPrimaryColumnIndex(columnBase, col)
         );
         return reader.getColumn(absoluteColumnIndex).getChar(offset);
+    }
+
+    @Override
+    public DirectSequence getDirectStr(int col) {
+        final long recordIndex = getAdjustedRecordIndex(col) * Long.BYTES;
+        final int absoluteColumnIndex = ifOffsetNegThen0ElseValue(
+                recordIndex,
+                TableReader.getPrimaryColumnIndex(columnBase, col)
+        );
+        return reader.getColumn(absoluteColumnIndex).getDirectStr(
+                reader.getColumn(absoluteColumnIndex + 1).getLong(recordIndex)
+        );
     }
 
     @Override