--- a/core/src/main/java/io/questdb/cairo/TableReaderSelectedColumnRecord.java
+++ b/core/src/main/java/io/questdb/cairo/TableReaderSelectedColumnRecord.java
@@ -27,6 +27,7 @@ package io.questdb.cairo;
 import io.questdb.cairo.sql.Record;
 import io.questdb.std.*;
 import io.questdb.std.str.CharSink;
+import io.questdb.std.str.DirectSequence;
 import org.jetbrains.annotations.NotNull;
 
 import static io.questdb.cairo.TableReaderRecord.ifOffsetNegThen0ElseValue;
@@ -97,6 +98,19 @@ public class TableReaderSelectedColumnRecord implements Record {
         final long offset = getAdjustedRecordIndex(col) * Character.BYTES;
         final int absoluteColumnIndex = ifOffsetNegThen0ElseValue(offset, index);
         return reader.getColumn(absoluteColumnIndex).getChar(offset);
+    }
+
+    @Override
+    public DirectSequence getDirectStr(int columnIndex) {
+        final int col = deferenceColumn(columnIndex);
+        final long recordIndex = getAdjustedRecordIndex(col) * Long.BYTES;
+        final int absoluteColumnIndex = ifOffsetNegThen0ElseValue(
+                recordIndex,
+                TableReader.getPrimaryColumnIndex(columnBase, col)
+        );
+        long offset = reader.getColumn(absoluteColumnIndex + 1).getLong(recordIndex);
+        assert recordIndex != 0 || (offset == 0 || offset == Numbers.LONG_NaN);
+        return reader.getColumn(absoluteColumnIndex).getDirectStr(offset);
     }
 
     @Override