--- a/core/src/main/java/io/questdb/cairo/sql/Record.java
+++ b/core/src/main/java/io/questdb/cairo/sql/Record.java
@@ -26,8 +26,9 @@ package io.questdb.cairo.sql;
 
 import io.questdb.std.BinarySequence;
 import io.questdb.std.Long256;
-import io.questdb.std.str.Utf16Sink;
 import io.questdb.std.str.CharSink;
+import io.questdb.std.str.DirectSequence;
+import io.questdb.std.str.Utf16Sink;
 
 /**
  * Access the value of columns of a table record by column index.
@@ -100,6 +101,16 @@ public interface Record {
      */
     default long getDate(int col) {
         return getLong(col);
+    }
+
+    /**
+     * Gets the off-heap (direct) value of a string column by index
+     *
+     * @param col numeric index of the column
+     * @return direct string, null if string is empty
+     */
+    default DirectSequence getDirectStr(int col) {
+        throw new UnsupportedOperationException();
     }
 
     /**
@@ -241,6 +252,7 @@ public interface Record {
      * @param col numeric index of the column
      * @return 64-bit integer
      */
+    @SuppressWarnings("unused")
     default long getLongIPv4(int col) {
         throw new UnsupportedOperationException();
     }