--- a/core/src/main/java/io/questdb/cairo/vm/NullMemoryMR.java
+++ b/core/src/main/java/io/questdb/cairo/vm/NullMemoryMR.java
@@ -28,7 +28,7 @@ import io.questdb.cairo.TableUtils;
 import io.questdb.cairo.vm.api.MemoryMR;
 import io.questdb.std.*;
 import io.questdb.std.str.CharSink;
-import io.questdb.std.str.DirectSequence;
+import io.questdb.std.str.DirectCharSequence;
 import io.questdb.std.str.LPSZ;
 
 public class NullMemoryMR implements MemoryMR {
@@ -74,7 +74,7 @@ public class NullMemoryMR implements MemoryMR {
     }
 
     @Override
-    public DirectSequence getDirectStr(long offset) {
+    public DirectCharSequence getDirectStr(long offset) {
         return null;
     }