--- a/core/src/main/java/io/questdb/cairo/vm/api/MemoryR.java
+++ b/core/src/main/java/io/questdb/cairo/vm/api/MemoryR.java
@@ -29,6 +29,7 @@ import io.questdb.std.Long256;
 import io.questdb.std.Long256Acceptor;
 import io.questdb.std.Unsafe;
 import io.questdb.std.str.CharSink;
+import io.questdb.std.str.DirectSequence;
 
 import java.io.Closeable;
 
@@ -51,6 +52,14 @@ public interface MemoryR extends Closeable {
     byte getByte(long offset);
 
     char getChar(long offset);
+
+    /**
+     * Returns UTF-16 encoded off-heap string.
+     * <p>
+     * Must return off-heap strings with stable pointers, i.e. once a string is returned,
+     * its pointer remains actual until the memory is closed.
+     */
+    DirectSequence getDirectStr(long offset);
 
     double getDouble(long offset);