--- a/core/src/main/java/io/questdb/cairo/vm/api/NullMemory.java
+++ b/core/src/main/java/io/questdb/cairo/vm/api/NullMemory.java
@@ -28,7 +28,7 @@ import io.questdb.std.BinarySequence;
 import io.questdb.std.FilesFacade;
 import io.questdb.std.Long256;
 import io.questdb.std.Long256Acceptor;
-import io.questdb.std.str.DirectSequence;
+import io.questdb.std.str.DirectCharSequence;
 import io.questdb.std.str.LPSZ;
 import org.jetbrains.annotations.NotNull;
 
@@ -74,7 +74,7 @@ public class NullMemory implements MemoryMAR, MemoryCARW {
     }
 
     @Override
-    public DirectSequence getDirectStr(long offset) {
+    public DirectCharSequence getDirectStr(long offset) {
         throw new UnsupportedOperationException();
     }