--- a/core/src/main/java/io/questdb/griffin/SqlUtil.java
+++ b/core/src/main/java/io/questdb/griffin/SqlUtil.java
@@ -653,7 +653,7 @@ public class SqlUtil {
 
     public static boolean isParallelismSupported(ObjList<Function> functions) {
         for (int i = 0, n = functions.size(); i < n; i++) {
-            if (!functions.getQuick(i).isParallelismSupported()) {
+            if (!functions.getQuick(i).supportsParallelism()) {
                 return false;
             }
         }