--- a/core/src/main/java/io/questdb/griffin/engine/functions/MultiArgFunction.java
+++ b/core/src/main/java/io/questdb/griffin/engine/functions/MultiArgFunction.java
@@ -64,18 +64,6 @@ public interface MultiArgFunction extends Function {
     }
 
     @Override
-    default boolean isParallelismSupported() {
-        final ObjList<Function> args = getArgs();
-        for (int i = 0, n = args.size(); i < n; i++) {
-            final Function function = args.getQuick(i);
-            if (!function.isParallelismSupported()) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    @Override
     default boolean isReadThreadSafe() {
         final ObjList<Function> args = getArgs();
         for (int i = 0, n = args.size(); i < n; i++) {
@@ -93,6 +81,18 @@ public interface MultiArgFunction extends Function {
         for (int i = 0, n = args.size(); i < n; i++) {
             final Function function = args.getQuick(i);
             if (!function.isRuntimeConstant() && !function.isConstant()) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    @Override
+    default boolean supportsParallelism() {
+        final ObjList<Function> args = getArgs();
+        for (int i = 0, n = args.size(); i < n; i++) {
+            final Function function = args.getQuick(i);
+            if (!function.supportsParallelism()) {
                 return false;
             }
         }