… core/src/main/java/io/questdb/griffin/engine/functions/QuaternaryFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/SymbolFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/TernaryFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/UnaryFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/columns/StrColumn.java
… core/src/main/java/io/questdb/griffin/engine/functions/columns/SymbolColumn.java
… core/src/main/java/io/questdb/griffin/engine/functions/constants/SymbolConstant.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/AbstractCountGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/AbstractCovarGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/AbstractStdDevGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/ApproxCountDistinctIPv4GroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/ApproxCountDistinctIntGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/ApproxCountDistinctLongGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/ApproxPercentileDoubleGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/ApproxPercentileDoublePackedGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/ApproxPercentileLongGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/ApproxPercentileLongPackedGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/AvgDoubleGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/CorrGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/CountDistinctIPv4GroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/CountDistinctIntGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/CountDistinctLong256GroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/CountDistinctLongGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/CountDistinctStringGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/CountDistinctSymbolGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/CountDistinctUuidGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/CountLongConstGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/FirstBooleanGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/FirstByteGroupByFunction.java
… core/src/main/java/io/questdb/griffin/engine/functions/groupby/FirstCharGroupByFunction.java