Introduces `DirectCharSequence getDirectStr()` method available only on string columns. The method is guaranteed to return off-heap strings with stable pointers, i.e. once a string is returned, its pointer remains actual until the end of query execution. Introduces `FirstDirectStrGroupByFunction` and other first/last SQL functions that don't materialize string value in a sink, but instead store `DirectCharSequence`'s pointer. This way last(str)/last_not_null(str) functions don't have to copy each row's string values into an off-heap sink.

Also, fixes writing an empty string to an empty `GroupByCharSink`.

Local benchmark with ClickBench data:
```sql
-- master 2.98s
-- patch 1.19s
select date_trunc('minute', eventtime) eventtime, count(), first(title), last(title)
from hits;

-- master 425ms
-- patch 430ms
select date_trunc('minute', eventtime) eventtime, first(title)
from hits;

-- master 1.97s
-- patch 970ms
select date_trunc('minute', eventtime) eventtime, last(title)
from hits;
```